common: sync debug functions from WINE
To get rid of -Wformat warnings.

+++ common/debug.h
@@ -174,7 +174,7 @@ static inline const char *nine_dbgstr_guid( const GUID *id )
     if (!id) return "(null)";
     if (!((ULONG_PTR)id >> 16)) return nine_dbg_sprintf( "<guid-0x%04hx>", (WORD)(ULONG_PTR)id );
     return nine_dbg_sprintf( "{%08x-%04x-%04x-%02x%02x-%02x%02x%02x%02x%02x%02x}",
-                             id->Data1, id->Data2, id->Data3,
+                             (unsigned int)id->Data1, id->Data2, id->Data3,
                              id->Data4[0], id->Data4[1], id->Data4[2], id->Data4[3],
                              id->Data4[4], id->Data4[5], id->Data4[6], id->Data4[7] );
 }
@@ -183,15 +183,15 @@ static inline const char *nine_dbgstr_guid( const GUID *id )
 static inline const char *nine_dbgstr_point( const POINT *pt )
 {
     if (!pt) return "(null)";
-    return nine_dbg_sprintf( "(%d,%d)", pt->x, pt->y );
+    return nine_dbg_sprintf( "(%d,%d)", (int)pt->x, (int)pt->y );
 }
 
 /* see WINE's wine_dbgstr_rect() */
 static inline const char *nine_dbgstr_rect( const RECT *rect )
 {
     if (!rect) return "(null)";
-    return nine_dbg_sprintf( "(%d,%d)-(%d,%d)", rect->left, rect->top,
-                             rect->right, rect->bottom );
+    return nine_dbg_sprintf( "(%d,%d)-(%d,%d)", (int)rect->left, (int)rect->top,
+                             (int)rect->right, (int)rect->bottom );
 }
 
 #endif /* __COMMON_DEBUG_H */
